Split the widget extension sources per type and documented them
Follow-up to cf4cea5.

Widgets.swift held four types, against the one-type-per-file rule in AGENTS.md, and none of them carried a documentation comment. Each now has its own file and the documentation the rest of the project's sources carry, saying what the type is for and — since these are scaffolding — that a real widget replaces them rather than grows out of them.

Renamed away from the template's names while splitting: Provider and SimpleEntry said nothing as module-level types, and struct Widgets shared its name with the module. The Placeholder prefix states what they are.

Also drops a force unwrap on a decorative date, marks the two diagnostic labels Text(verbatim:) so they stop being extracted into a catalog nobody will translate them into, and removes the commented-out relevances() stub. The widget's own display name and description stay as the template wrote them, since they are the strings the real widget replaces.
